requirements-ini.txt
absl-py            2.0.0
certifi            2022.12.7
charset-normalizer 3.3.2
colorama           0.4.6
cycler             0.11.0
fonttools          4.38.0
glfw               2.6.4
idna               3.6
importlib-metadata 6.7.0
Jinja2             3.1.3
joblib             1.3.2
kiwisolver         1.4.5
llvmlite           0.39.1
MarkupSafe         2.1.5
matplotlib         3.5.3
mujoco             2.3.6
networkx           2.6.3
numba              0.56.4
numpy              1.21.6
numpy-quaternion   2022.4.2
opencv-python      4.9.0.80
packaging          23.2
Pillow             9.5.0
pip                22.3.1
psutil             5.9.8
pydot              1.4.2
pygraphviz         1.7
PyOpenGL           3.1.7
pyparsing          3.1.1
python-dateutil    2.8.2
requests           2.31.0
robosuite          1.4.0
scikit-learn       1.0.2
scipy              1.7.3
setuptools         65.6.3
six                1.16.0
termcolor          2.3.0
threadpoolctl      3.1.0
torch              1.12.0+cpu
torch-cluster      1.6.1
torch-geometric    2.3.1
torch-scatter      2.1.1
torch-sparse       0.6.17
torch-spline-conv  1.2.2
tqdm               4.66.2
typing_extensions  4.7.1
urllib3            2.0.7
wheel              0.38.4
wincertstore       0.2
xmltodict          0.13.0
zipp               3.15.0